requirements.txt
PyYAML==6.0.2
aiogram==3.17.0
MetaTrader5==5.0.45
numpy==1.26.4
